requirements.txt
openpyxl
python-dateutil
cx-Oracle
et-xmlfile
six